gcp-functions/vip-users/requirements.txt
functions-framework==3.*
kafka-python==2.0.2
